requirements.txt
numpy~=1.24.1
torch~=1.13.1
Pillow~=9.3.0